add send byte api

+++ xquic/src/main/cpp/main/include/xquic_h3_ctrl.h
@@ -14,7 +14,7 @@ extern "C"{
 
 ssize_t client_send_h3_content(xqc_cli_user_stream_t *user_stream);
 
-int client_send_h3_requests(xqc_cli_user_conn_t *user_conn,
+ssize_t client_send_h3_requests(xqc_cli_user_conn_t *user_conn,
                             xqc_cli_user_stream_t *user_stream, xqc_cli_request_t *req);
 
 ssize_t client_send_H3_ping(xqc_cli_user_conn_t *user_conn, char *ping_user_data);
